Make release.sh smarter, fix #1043.

+++ release.sh
@@ -20,7 +20,7 @@ if [ $# = 0 ]; then
     echo "Usage: $0 <version>"
     echo
     echo "  This is a release script for Gittip. We bump the version number in "
-    echo "  gittip/__init__.py and then do a git dance, pushing to Heroku."
+    echo "  www/version.txt and then do a git dance, pushing to Heroku."
     echo
     exit
 fi
@@ -60,48 +60,58 @@ if [ $1 ]; then
         exit
     fi
 
+    if [ "`git rev-parse --abbrev-ref HEAD`" != "master" ]; then
+        echo "Not on master, checkout master first."
+        exit
+    fi
+
+    # Make sure we have the latest master.
+    # ====================================
+
+    git pull
+
     if [ "`git tag | grep $1`" ]; then
         echo "Version $1 is already git tagged."
-    else
-        confirm "Tag and push version $1?"
-        if [ $? -eq 0 ]; then
+        exit
+    fi
 
-            # Bump the version.
-            # =================
+    if ! grep -e \-dev$ www/version.txt > /dev/null; then
+        echo "Current version does not end with '-dev'."
+        exit
+    fi
 
-            printf "$1\n" > www/version.txt
-            git commit www/version.txt -m"Bump version to $1"
-            git tag $1
+    confirm "Tag and push version $1?"
+    if [ $? -eq 0 ]; then
 
+        # Bump the version.
+        # =================
 
-            # Deploy to Heroku.
-            # =================
-            # If this fails we still want to bump the version again, so modify
-            # bash error handling around this call.
-
-            set +e
-            git push heroku master
-            set -e
+        printf "$1\n" > www/version.txt
+        git commit www/version.txt -m"Bump version to $1"
+        git tag $1
 
 
-            # Bump the version again.
-            # =======================
-            # We're using a Pythonic convention here by using -dev to mean, "a
-            # dev version following $whatever." We escape the dash for bash's
-            # sake
+        # Deploy to Heroku.
+        # =================
 
-            printf "$1\055dev\n" > www/version.txt
-            git commit www/version.txt -m"Bump version to $1-dev"
+        git push heroku master
 
 
-            # Push to GitHub.
-            # ===============
-            # This will break if we didn't pull before releasing. We should do
-            # that.
+        # Bump the version again.
+        # =======================
+        # We're using a Pythonic convention here by using -dev to mean, "a
+        # dev version following $whatever." We escape the dash for bash's
+        # sake
 
-            git push
-            git push --tags
+        printf "$1\055dev\n" > www/version.txt
+        git commit www/version.txt -m"Bump version to $1-dev"
 
-        fi
+
+        # Push to GitHub.
+        # ===============
+
+        git push
+        git push --tags
+
     fi
 fi
